refactor(server): split remaining websocket tests
Keep the WebSocket route test parent focused on shared fixtures and move the remaining monolithic test groups into child modules by concern: frame parsing, MUC, IQ, messages, broadcast, stream features, and miscellaneous service checks. This preserves the existing test bodies and keeps issue #350 cleanup test-only.

Constraint: PR is test-only; no XMPP wire shape, XEP behavior, or production WebSocket runtime changes

Constraint: Raw XML fixtures are preserved only where the moved tests already used them as parser/protocol inputs

Rejected: Move helpers into a child module | existing split modules depend on parent-level fixtures

Rejected: Production route refactors in this PR | review surface must stay limited to test organization

Confidence: high

Scope-risk: narrow

Directive: Keep websocket/tests.rs as shared test setup; add behavior-specific route tests under websocket/tests/*.rs

Tested: cargo test -p waddle-server server::routes::websocket

Tested: cargo test -p waddle-server server::routes::websocket::tests::frame_parsing

Tested: cargo test -p waddle-server server::routes::websocket::tests::muc

Tested: cargo test -p waddle-server server::routes::websocket::tests::iq

Tested: cargo test -p waddle-server server::routes::websocket::tests::messages

Tested: cargo test -p waddle-server server::routes::websocket::tests::broadcast

Tested: cargo test -p waddle-server server::routes::websocket::tests::misc

Tested: cargo test -p waddle-server server::routes::websocket::tests::stream_features

Tested: cargo test -p waddle-server

Tested: cargo clippy -p waddle-server --tests -- -D warnings

Tested: git diff --check

Tested: adversarial review agents found no actionable findings
